Refactor data loaders to use Spanish naming convention
